Check for admin role for any endpoints that change data

+++ src/Controller/CategoryController.php
@@ -1,5 +1,4 @@
 <?php
-
 namespace App\Controller;
 
 use App\Entity\Category;
@@ -12,11 +11,14 @@ use App\Repository\CreationTypeRepository;
 use App\Repository\UserRepository;
 
 /**
+ *
  * @Route("/category")
  */
 class CategoryController extends AbstractController
 {
+
     /**
+     *
      * @Route("/", name="category_index", methods={"GET"})
      */
     public function index(): Response
@@ -26,22 +28,26 @@ class CategoryController extends AbstractController
             ->findAll();
 
         return $this->render('category/index.html.twig', [
-            'categories' => $categories,
+            'categories' => $categories
         ]);
     }
 
     /**
+     *
      * @Route("/new", name="category_new", methods={"GET","POST"})
      */
     public function new(Request $request, CreationTypeRepository $creationTypeRespository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         $category = new Category();
         $form = $this->createForm(CategoryType::class, $category);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $entityManager = $this->getDoctrine()->getManager();
-            $category->setCreationType($creationTypeRespository->findByName('Manual'))->setCreatedBy($this->getUser());
+            $category->setCreationType($creationTypeRespository->findByName('Manual'))
+                ->setCreatedBy($this->getUser());
             $entityManager->persist($category);
             $entityManager->flush();
 
@@ -51,11 +57,12 @@ class CategoryController extends AbstractController
 
         return $this->render('category/new.html.twig', [
             'category' => $category,
-            'form' => $form->createView(),
+            'form' => $form->createView()
         ]);
     }
-    
+
     /**
+     *
      * @Route("/export", name="category_export")
      */
     public function export(): Response
@@ -64,46 +71,44 @@ class CategoryController extends AbstractController
     }
 
     /**
-     * @Route("/{id}", name="category_show", methods={"GET"})
-     */
-    public function show(Category $category): Response
-    {
-        return $this->render('category/show.html.twig', [
-            'category' => $category,
-        ]);
-    }
-
-    /**
+     *
      * @Route("/{id}/edit", name="category_edit", methods={"GET","POST"})
      */
     public function edit(Request $request, Category $category): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         $form = $this->createForm(CategoryType::class, $category);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
+            $this->getDoctrine()
+                ->getManager()
+                ->flush();
 
             return $this->redirectToRoute('category_index');
         }
 
         return $this->render('category/edit.html.twig', [
             'category' => $category,
-            'form' => $form->createView(),
+            'form' => $form->createView()
         ]);
     }
 
     /**
+     *
      * @Route("/{id}", name="category_delete", methods={"DELETE"})
      */
     public function delete(Request $request, Category $category): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$category->getId(), $request->request->get('_token'))) {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
+        if ($this->isCsrfTokenValid('delete' . $category->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($category);
             $entityManager->flush();
-            
-            $this->addFlash('success',sprintf('Category %s Deleted', $category->getName()));
+
+            $this->addFlash('success', sprintf('Category %s Deleted', $category->getName()));
         }
 
         return $this->redirectToRoute('category_index');
