+++ src/Controller/SslAccountController.php
@@ -1,5 +1,4 @@
 <?php
-
 namespace App\Controller;
 
 use App\Entity\SslAccount;
@@ -10,11 +9,14 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
 /**
+ *
  * @Route("/ssl/account")
  */
 class SslAccountController extends AbstractController
 {
+
     /**
+     *
      * @Route("/", name="ssl_account_index", methods={"GET"})
      */
     public function index(): Response
@@ -24,15 +26,18 @@ class SslAccountController extends AbstractController
             ->findAll();
 
         return $this->render('ssl_account/index.html.twig', [
-            'ssl_accounts' => $sslAccounts,
+            'ssl_accounts' => $sslAccounts
         ]);
     }
 
     /**
+     *
      * @Route("/new", name="ssl_account_new", methods={"GET","POST"})
      */
     public function new(Request $request): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         $sslAccount = new SslAccount();
         $form = $this->createForm(SslAccountType::class, $sslAccount);
         $form->handleRequest($request);
@@ -47,46 +52,44 @@ class SslAccountController extends AbstractController
 
         return $this->render('ssl_account/new.html.twig', [
             'ssl_account' => $sslAccount,
-            'form' => $form->createView(),
+            'form' => $form->createView()
         ]);
     }
 
     /**
-     * @Route("/{id}", name="ssl_account_show", methods={"GET"})
-     */
-    public function show(SslAccount $sslAccount): Response
-    {
-        return $this->render('ssl_account/show.html.twig', [
-            'ssl_account' => $sslAccount,
-        ]);
-    }
-
-    /**
+     *
      * @Route("/{id}/edit", name="ssl_account_edit", methods={"GET","POST"})
      */
     public function edit(Request $request, SslAccount $sslAccount): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         $form = $this->createForm(SslAccountType::class, $sslAccount);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
+            $this->getDoctrine()
+                ->getManager()
+                ->flush();
 
             return $this->redirectToRoute('ssl_account_index');
         }
 
         return $this->render('ssl_account/edit.html.twig', [
             'ssl_account' => $sslAccount,
-            'form' => $form->createView(),
+            'form' => $form->createView()
         ]);
     }
 
     /**
+     *
      * @Route("/{id}", name="ssl_account_delete", methods={"DELETE"})
      */
     public function delete(Request $request, SslAccount $sslAccount): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$sslAccount->getId(), $request->request->get('_token'))) {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
+        if ($this->isCsrfTokenValid('delete' . $sslAccount->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($sslAccount);
             $entityManager->flush();
