+++ src/Controller/SslCertTypeController.php
@@ -1,5 +1,4 @@
 <?php
-
 namespace App\Controller;
 
 use App\Entity\SslCertType;
@@ -10,11 +9,14 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
 /**
+ *
  * @Route("/ssl/type")
  */
 class SslCertTypeController extends AbstractController
 {
+
     /**
+     *
      * @Route("/", name="ssl_cert_type_index", methods={"GET"})
      */
     public function index(): Response
@@ -24,15 +26,18 @@ class SslCertTypeController extends AbstractController
             ->findAll();
 
         return $this->render('ssl_cert_type/index.html.twig', [
-            'ssl_cert_types' => $sslCertTypes,
+            'ssl_cert_types' => $sslCertTypes
         ]);
     }
 
     /**
+     *
      * @Route("/new", name="ssl_cert_type_new", methods={"GET","POST"})
      */
     public function new(Request $request): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         $sslCertType = new SslCertType();
         $form = $this->createForm(SslCertTypeType::class, $sslCertType);
         $form->handleRequest($request);
@@ -48,11 +53,12 @@ class SslCertTypeController extends AbstractController
 
         return $this->render('ssl_cert_type/new.html.twig', [
             'ssl_cert_type' => $sslCertType,
-            'form' => $form->createView(),
+            'form' => $form->createView()
         ]);
     }
-    
+
     /**
+     *
      * @Route("/export", name="ssl_cert_type_export")
      */
     public function export(): Response
@@ -61,25 +67,20 @@ class SslCertTypeController extends AbstractController
     }
 
     /**
-     * @Route("/{id}", name="ssl_cert_type_show", methods={"GET"})
-     */
-    public function show(SslCertType $sslCertType): Response
-    {
-        return $this->render('ssl_cert_type/show.html.twig', [
-            'ssl_cert_type' => $sslCertType,
-        ]);
-    }
-
-    /**
+     *
      * @Route("/{id}/edit", name="ssl_cert_type_edit", methods={"GET","POST"})
      */
     public function edit(Request $request, SslCertType $sslCertType): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         $form = $this->createForm(SslCertTypeType::class, $sslCertType);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
+            $this->getDoctrine()
+                ->getManager()
+                ->flush();
 
             $this->addFlash('success', sprintf('SSL Type %s Updated', $sslCertType->getType()));
             return $this->redirectToRoute('ssl_cert_type_index');
@@ -87,16 +88,19 @@ class SslCertTypeController extends AbstractController
 
         return $this->render('ssl_cert_type/edit.html.twig', [
             'ssl_cert_type' => $sslCertType,
-            'form' => $form->createView(),
+            'form' => $form->createView()
         ]);
     }
 
     /**
+     *
      * @Route("/{id}", name="ssl_cert_type_delete", methods={"DELETE"})
      */
     public function delete(Request $request, SslCertType $sslCertType): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$sslCertType->getId(), $request->request->get('_token'))) {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
+        if ($this->isCsrfTokenValid('delete' . $sslCertType->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($sslCertType);
             $entityManager->flush();
